+++ crates/webz-keys/Cargo.toml
@@ -19,11 +19,15 @@ webz-common = { path = "../webz-common" }
 js-sys.workspace = true
 thiserror.workspace = true
 wasm-bindgen.workspace = true
-zcash_primitives.workspace = true
+zcash_primitives = { workspace = true, features = ["transparent-inputs"] }
 zcash_keys.workspace = true
 bip0039.workspace = true
 zip32 = "0.1"
-sapling = { workspace = true}
+sapling = { workspace = true }
+bip32 = "0.5"
+pczt = { workspace = true, default-features = false, features = ["signer", "orchard", "sapling", "transparent", "tx-extractor"] }
+orchard = { version = "0.10.1", default-features = false }
+wasm-bindgen-futures = "0.4.43"
 
 # fixes "failed to resolve: use of undeclared crate or module `imp`" error
 getrandom = { version = "0.2", features = ["js"] }
