+++ crates/webz-keys/src/lib.rs
@@ -3,6 +3,8 @@
 
 mod error;
 mod keys;
+mod pczt_sign;
 
 pub use error::*;
 pub use keys::*;
+pub use pczt_sign::*;
