+++ crates/webz-wallet/src/bindgen/mod.rs
@@ -1,5 +1,4 @@
 use serde::{Deserialize, Serialize};
 
-mod pczt;
 pub mod proposal;
 pub mod wallet;
