+++ crates/webz-wallet/src/bindgen/wallet.rs
@@ -10,12 +10,9 @@ use tonic_web_wasm_client::Client;
 
 use crate::error::Error;
 use crate::wallet::usk_from_seed_str;
-use crate::{
-    bindgen::{pczt::Pczt, proposal::Proposal},
-    Wallet, PRUNING_DEPTH,
-};
+use crate::{bindgen::proposal::Proposal, Wallet, PRUNING_DEPTH};
 use wasm_thread as thread;
-use webz_common::Network;
+use webz_common::{Network, Pczt};
 use webz_keys::{ProofGenerationKey, SeedFingerprint, UnifiedSpendingKey};
 use zcash_address::ZcashAddress;
 use zcash_client_backend::data_api::{InputSource, WalletRead};
@@ -433,27 +430,6 @@ impl WebWallet {
     ) -> Result<Pczt, Error> {
         self.inner
             .pczt_prove(pczt.into(), sapling_proof_gen_key.map(Into::into))
-            .await
-            .map(Into::into)
-    }
-
-    /// Signs and applies signatures to a PCZT.
-    /// Should in a secure environment (e.g. Metamask snap).
-    ///
-    /// # Arguments
-    ///
-    /// * `pczt` - The PCZT that needs to signed
-    /// * `usk` - UnifiedSpendingKey used to sign the PCZT
-    /// * `seed_fp` - The fingerprint of the seed used to create `usk`
-    ///
-    pub async fn pczt_sign(
-        &self,
-        pczt: Pczt,
-        usk: UnifiedSpendingKey,
-        seed_fp: SeedFingerprint,
-    ) -> Result<Pczt, Error> {
-        self.inner
-            .pczt_sign(pczt.into(), usk.into(), seed_fp.into())
             .await
             .map(Into::into)
     }
