+++ crates/webz-wallet/src/wallet.rs
@@ -549,156 +549,6 @@ where
         Ok(pczt)
     }
 
-    pub async fn pczt_sign(
-        &self,
-        pczt: Pczt,
-        usk: UnifiedSpendingKey,
-        seed_fp: SeedFingerprint,
-    ) -> Result<Pczt, Error> {
-        // Find all the spends matching our seed.
-        enum KeyRef {
-            Orchard {
-                index: usize,
-            },
-            Sapling {
-                index: usize,
-            },
-            Transparent {
-                index: usize,
-                scope: TransparentKeyScope,
-                address_index: NonHardenedChildIndex,
-            },
-        }
-        let mut keys = BTreeMap::<zip32::AccountId, Vec<KeyRef>>::new();
-        let pczt = Verifier::new(pczt)
-            .with_orchard::<Infallible, _>(|bundle| {
-                for (index, action) in bundle.actions().iter().enumerate() {
-                    if let Some(account_index) = action
-                        .spend()
-                        .zip32_derivation()
-                        .as_ref()
-                        .and_then(|derivation| {
-                            derivation.extract_account_index(
-                                &seed_fp,
-                                zip32::ChildIndex::hardened(
-                                    self.network.network_type().coin_type(),
-                                ),
-                            )
-                        })
-                    {
-                        keys.entry(account_index)
-                            .or_default()
-                            .push(KeyRef::Orchard { index });
-                    }
-                }
-                Ok(())
-            })
-            .map_err(|e| Error::PcztSign(format!("Invalid PCZT: {:?}", e)))?
-            .with_sapling::<Infallible, _>(|bundle| {
-                for (index, spend) in bundle.spends().iter().enumerate() {
-                    if let Some(account_index) =
-                        spend.zip32_derivation().as_ref().and_then(|derivation| {
-                            derivation.extract_account_index(
-                                &seed_fp,
-                                zip32::ChildIndex::hardened(
-                                    self.network.network_type().coin_type(),
-                                ),
-                            )
-                        })
-                    {
-                        keys.entry(account_index)
-                            .or_default()
-                            .push(KeyRef::Sapling { index });
-                    }
-                }
-                Ok(())
-            })
-            .map_err(|e| Error::PcztSign(format!("Invalid PCZT: {:?}", e)))?
-            .with_transparent::<Infallible, _>(|bundle| {
-                for (index, input) in bundle.inputs().iter().enumerate() {
-                    for derivation in input.bip32_derivation().values() {
-                        if let Some((account_index, scope, address_index)) = derivation
-                            .extract_bip_44_fields(
-                                &seed_fp,
-                                bip32::ChildNumber(
-                                    self.network.network_type().coin_type()
-                                        | bip32::ChildNumber::HARDENED_FLAG,
-                                ),
-                            )
-                        {
-                            keys.entry(account_index)
-                                .or_default()
-                                .push(KeyRef::Transparent {
-                                    index,
-                                    scope,
-                                    address_index,
-                                });
-                        }
-                    }
-                }
-                Ok(())
-            })
-            .map_err(|e| Error::PcztSign(format!("Invalid PCZT: {:?}", e)))?
-            .finish();
-
-        let mut signer = Signer::new(pczt).unwrap();
-        //.map_err(|e| anyhow!("Failed to initialize Signer: {:?}", e))?;
-        for (account_index, spends) in keys {
-            // let usk = UnifiedSpendingKey::from_seed(&params, seed, account_index)?;
-            for keyref in spends {
-                match keyref {
-                    KeyRef::Orchard { index } => {
-                        signer
-                            .sign_orchard(
-                                index,
-                                &orchard::keys::SpendAuthorizingKey::from(usk.orchard()),
-                            )
-                            .map_err(|e| {
-                                Error::PcztSign(format!(
-                                    "Failed to sign Orchard spend {index}: {:?}",
-                                    e
-                                ))
-                            })?;
-                    }
-                    KeyRef::Sapling { index } => {
-                        signer
-                            .sign_sapling(index, &usk.sapling().expsk.ask)
-                            .map_err(|e| {
-                                Error::PcztSign(format!(
-                                    "Failed to sign Sapling spend {index}: {:?}",
-                                    e
-                                ))
-                            })?;
-                    }
-                    KeyRef::Transparent {
-                        index,
-                        scope,
-                        address_index,
-                    } => signer
-                        .sign_transparent(
-                            index,
-                            &usk.transparent()
-                                .derive_secret_key(scope, address_index)
-                                .map_err(|e| {
-                                    Error::PcztSign(format!(
-                                        "Failed to derive transparent key at .../{:?}/{:?}: {:?}",
-                                        scope, address_index, e,
-                                    ))
-                                })?,
-                        )
-                        .map_err(|e| {
-                            Error::PcztSign(format!(
-                                "Failed to sign transparent input {index}: {:?}",
-                                e
-                            ))
-                        })?,
-                }
-            }
-        }
-
-        Ok(signer.finish())
-    }
-
     pub async fn pczt_send(&self, pczt: Pczt) -> Result<(), Error> {
         let prover = LocalTxProver::bundled();
         let (spend_vk, output_vk) = prover.verifying_keys();
@@ -716,7 +566,6 @@ where
                 e
             ))
         })?;
-
         drop(db);
         self.send_authorized_transactions(&NonEmpty::new(txid))
             .await
